+++ src/components/ProfileInformation/ProfileInformation.tsx
@@ -10,8 +10,8 @@ import Share from '../../assets/icons/Share.svg'
 import Twitter from '../../assets/icons/Twitter.svg'
 import Wallet from '../../assets/icons/Wallet.svg'
 import { config } from '../../modules/config/config'
-import copyText from '../../utils/copyText'
-import { useTimer } from '../../utils/timer'
+import copyText from '../../utils/useCopy'
+import { useTimer } from '../../utils/useTimer'
 import { EDIT_PROFILE_URL } from '../Avatar/consts'
 import { twitterURL } from './contsts'
 import { Props } from './ProfileInformation.types'
