+++ src/utils/useTimer.ts
